requirements.txt
streamlit>=1.31.0
google-genai>=0.1.0
python-dotenv>=1.0.0
pypdf>=4.0.0